Add files via upload
This is an ArcGIS Toolbox that is associated with a directory of input data, which is too big for github, and is available upon request.  The most recent version of the toolbox as of Sept. 29, 2020. This toolbox needed to be compressed slightly,to fit below the 25 MB limit on github. The uncompressed version is available upon request.